% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dn.R
\name{add_dn_date}
\alias{add_dn_date}
\alias{cal_di}
\alias{dndate_start}
\title{add_dn_date}
\usage{
add_dn_date(img, date_begin = NULL, include.year = TRUE, chunksize = 8)

cal_di(date_begin, chunksize = 8)

dndate_start(year, di, chunksize = 8)
}
\description{
add_dn_date
}
\examples{
\dontrun{
dndate_start(2020, 46) \%>\% cal_di(include.year = FALSE)
}
}
