% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_buffer.R
\name{df2sf}
\alias{df2sf}
\title{df2sf}
\usage{
df2sf(d, coords = c("lon", "lat"), crs = 4326)
}
\description{
df2sf
}
\examples{
n <- 100
lon <- seq(1, 180, length.out = n)
lat <- seq(1, 90, length.out = n)

df <- data.frame(I = 1:n, i = 1:n, lon, lat)
df2sf(df)
}
\keyword{internal}
