% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{ee_aggregate}
\alias{ee_aggregate}
\alias{ee_aggregate_list}
\title{ee_aggregate}
\usage{
ee_aggregate(col, prop, reducerList = "mean", bandList = 0)

ee_aggregate_list(col, prop, reducerList = "mean", bandList = 0)
}
\description{
ee_aggregate
}
