% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_date.R
\name{ee_date}
\alias{ee_date}
\alias{as_date_img}
\alias{as_date_millis}
\alias{str_date}
\alias{ee_systemtime}
\alias{ee_yday}
\alias{ee_timestart}
\alias{ee_timeend}
\title{ee_date}
\usage{
as_date_img(x, prop = "system:time_start")

as_date_millis(x)

str_date(date)

ee_systemtime(x, prop = "system:time_start")

ee_yday(x, ...)

ee_timestart(x)

ee_timeend(x)
}
\arguments{
\item{prop}{"system:time_start" or "system:time_end"}
}
\value{
\itemize{
\item \code{ee_yday}: ee.Number
}
}
\description{
Get imgcol or img "system:time_start" or "system:time_end"
}
