% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_extract2.R
\name{ee_extract2}
\alias{ee_extract2}
\title{ee_extract2}
\usage{
ee_extract2(x, y, fun = ee$Reducer$mean(), scale = NULL, ...)
}
\arguments{
\item{x}{ee$Image.}

\item{y}{ee$Geometry$*, ee$Feature, ee$FeatureCollection, sfc or sf objects.}

\item{fun}{ee$Reducer object. Function to summarize the values. The function
must take a single numeric value as an argument and return a single value.
See details.}

\item{scale}{A nominal scale in meters of the Image projection to work in.
By default 1000.}

\item{...}{ee$Image$reduceRegions additional parameters. See
\code{ee_help(ee$Image$reduceRegions)} for more details.}
}
\description{
ee_extract2
}
