% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_export.R
\name{ee_get_proj}
\alias{ee_get_proj}
\title{Get projection of ee.Image or ee.ImageCollection}
\usage{
ee_get_proj(imgcol)
}
\arguments{
\item{imgcol}{ee.Image or ee.ImageCollection}
}
\value{
\itemize{
\item proj
\item scale
\item crs
\item transform
}
}
\description{
Get projection of ee.Image or ee.ImageCollection
}
\note{
After \code{map} operation, scale and transform information will disappear.
}
