% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_gif.R
\name{ee_gif}
\alias{ee_gif}
\title{ee_gif}
\usage{
ee_gif(
  imgcol,
  vis = NULL,
  mask = NULL,
  reigon = NULL,
  dimensions = 600,
  crs = "EPSG:3857",
  framesPerSecond = 2
)
}
\arguments{
\item{mask}{(Feature|Geometry|Object) The Geometry or Feature to clip to.}

\item{dimensions}{A number or pair of numbers in format c(WIDTH,HEIGHT).
Max dimensions of the thumbnail to render, in pixels. If only one number is passed,
it is used as the maximum, and the other dimension is computed by proportional scaling.}

\item{crs}{A CRS string specifying the projection of the output.
"EPSG:4326" and "EPSG:3857" (Web Mercator projection) are the most frequent used.}

\item{framesPerSecond}{String. Animation speed.}

\item{region}{(optional) ee$Geometry$Polygon, GeoJSON or c(E,S,W,N). Geospatial region
of the result. By default, the whole image.}

\item{crs_transform}{The affine transform to use for the output pixel grid.}

\item{scale}{A scale to determine the output pixel grid; ignored if both crs
and crs_transform are specified.}

\item{format}{String. The output format (only 'gif' is currently supported).}
}
\description{
ee_gif
}
