% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_date.R
\name{ee_systemtime}
\alias{ee_systemtime}
\alias{ee_timeStart}
\alias{ee_timeEnd}
\title{ee_systemtime}
\usage{
ee_systemtime(x, prop = "system:time_start")

ee_timeStart(x)

ee_timeEnd(x)
}
\arguments{
\item{prop}{"system:time_start" or "system:time_end"}
}
\description{
Get imgcol or img "system:time_start" or "system:time_end"
}
