% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{ee_tools}
\alias{ee_tools}
\alias{ee_getInfo}
\alias{ee_num2str}
\alias{getInfo}
\alias{getInfo2}
\alias{ee_aggregate_array}
\alias{ee_systemIndex}
\alias{ee_first}
\alias{ee_last}
\alias{ee_bandNames}
\alias{ee_propertyNames}
\alias{ee_properties}
\alias{ee_filterDate}
\alias{ee_filterBounds}
\alias{image_size}
\title{rgee \code{imgcol} tools collection}
\usage{
ee_getInfo(x)

ee_num2str(x)

getInfo(x)

getInfo2(value)

ee_aggregate_array(col, prop = "system:time_start")

ee_systemIndex(col)

ee_first(col)

ee_last(col)

ee_bandNames(x)

ee_propertyNames(col)

ee_properties(col, verbose = TRUE)

ee_filterDate(col, date_start, date_end = NULL)

ee_filterBounds(col, geom)

image_size(x)
}
\description{
rgee \code{imgcol} tools collection
}
