% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_export.R
\name{export_Img}
\alias{export_Img}
\title{export_Img}
\usage{
export_Img(img, task, options)
}
\description{
export_Img
}
\section{Some options to clip regional data}{
\itemize{
\item 
\enumerate{
\item crs + region + crsTransform    | √
}
\item 
\enumerate{
\item crs + region + dimensions      | √ (resample)
}
\item 
\enumerate{
\item crs + region + scale           | √
}
}
}

