% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{glue}
\alias{ee_Initialize}
\alias{ee}
\alias{Map}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{glue}{\code{\link[glue]{glue}}}

  \item{rgee}{\code{\link[rgee]{ee}}, \code{\link[rgee]{ee_Initialize}}, \code{\link[rgee]{Map}}}
}}

