% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgee2-package.R
\docType{package}
\name{rgee2-package}
\alias{rgee2}
\alias{rgee2-package}
\title{rgee2: Extra functions for R Google Earth Engine}
\description{
Google Earth Engine Extensions in R.
}
\author{
\strong{Maintainer}: Dongdondg Kong \email{kongdd.sysu@gmail.com} (\href{https://orcid.org/0000-0003-1836-8172}{ORCID})

}
\keyword{internal}
