% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_buffer.R
\name{st_point_buffer}
\alias{st_point_buffer}
\title{ee_point_buffer}
\usage{
st_point_buffer(st, scale = 500, half_win = 1)
}
\arguments{
\item{st}{data.table with the columns at least of \code{site}, \code{lon}, and \code{lat}.}

\item{scale}{in the unit of \code{m}}

\item{half_win}{half window of the buffer (in grids). window size equals to
\code{2*halfwin + 1}.}
}
\description{
ee_point_buffer
}
\examples{
\dontrun{
st <- st_212[, .(site, lon, lat, IGBP)]
st_point_buffer_3by3(st, 500)
}
}
