% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_init.R
\name{ee_init}
\alias{ee_init}
\title{ee_init}
\usage{
ee_init(drive = FALSE, ...)
}
\arguments{
\item{drive}{Logical (optional). If set to TRUE, the drive credential will be
cached in the path \code{~/.config/earthengine/}.}

\item{...}{others to \code{\link[rgee:ee_Initialize]{rgee::ee_Initialize()}}}
}
\description{
ee_init
}
\examples{
\dontrun{
# ee_Initialize(user = "cuijian426", drive = TRUE)
ee_init()
}
}
\seealso{
\code{\link[rgee:ee_Initialize]{rgee::ee_Initialize()}}
}
