% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{glue}
\alias{ee_Initialize}
\alias{ee}
\alias{Map}
\alias{make_date}
\alias{mutate}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{mutate}}}

  \item{glue}{\code{\link[glue]{glue}}}

  \item{lubridate}{\code{\link[lubridate:make_datetime]{make_date}}}

  \item{rgee}{\code{\link[rgee]{ee}}, \code{\link[rgee]{ee_Initialize}}, \code{\link[rgee]{Map}}}
}}

